+++ rapoc/models/__init__.py
@@ -1,3 +1,4 @@
 from .model import Model
 from .planck import Planck
 from .rosseland import Rosseland
+from .rayleigh import Rayleigh